Safari-609.1.11.1/LayoutTests/fast/css/url-with-multi-byte-unicode-escape.html

<!DOCTYPE html>
<html>
<head>
<style>#test { background-image: url("data:\100"); border-image: url(data:\100); } #test { background-color: green !important; }</style>
<script src="../../resources/js-test-pre.js"></script>
<script>
    if (window.testRunner)
        testRunner.dumpAsText();
</script>
</head>
<body>
<div id="test" style="width: 100px; height: 100px; background-color: red"></div>
<script>
    description('Test parsing a CSS URI containing a multi-byte Unicode escape sequence.');
    shouldBe('document.styleSheets[0].cssRules.length', '2');
    shouldBeEqualToString('document.styleSheets[0].cssRules[0].style.getPropertyValue("background-image")', 'url("data:%C4%80")');
    shouldBeEqualToString('document.styleSheets[0].cssRules[0].style.getPropertyValue("border-image-source")', 'url("data:%C4%80")');
    shouldBeEqualToString('document.styleSheets[0].cssRules[1].style.getPropertyValue("background-color")', 'green');
    shouldBeEqualToString('window.getComputedStyle(document.getElementById("test")).getPropertyValue("background-color")', 'rgb(0, 128, 0)');
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
