third_party/blink/web_tests/fast/css/url-with-multi-byte-unicode-escape-expected.txt

Test parsing a CSS URI containing a multi-byte Unicode escape sequence.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.styleSheets[0].cssRules.length is 2
PASS document.styleSheets[0].cssRules[0].style.getPropertyValue("background-image") is "url(\"data:Ā\")"
PASS document.styleSheets[0].cssRules[1].style.getPropertyValue("background-color") is "green"
PASS window.getComputedStyle(document.getElementById("test")).getPropertyValue("background-color") is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
